+++ A32NX/html_ui/Pages/VCockpit/Instruments/Airliners/A320_Neo/CDU/OPTIONS/A320_Neo_CDU_OPTIONS_AOC.js
@@ -1,0 +1,28 @@
+class CDU_OPTIONS_AOC {
+    static ShowPage(mcdu) {
+        mcdu.clearDisplay();
+
+        mcdu.setTemplate([
+            ["A32NX OPTIONS AOC"],
+            [""],
+            [""],
+            [""],
+            [""],
+            [""],
+            [""],
+            [""],
+            [""],
+            [""],
+            [""],
+            [""],
+            ["<RETURN[color]cyan"]
+        ]);
+
+        mcdu.leftInputDelay[5] = () => {
+            return mcdu.getDelaySwitchPage();
+        };
+        mcdu.onLeftInput[5] = () => {
+            CDU_OPTIONS_MainMenu.ShowPage(mcdu);
+        };
+    }
+}